Moved a couple of service tests to unit tests
