+++ services/github/github-directory-file-count.spec.js
@@ -1,0 +1,42 @@
+'use strict'
+
+const { expect } = require('chai')
+const { test, given } = require('sazerac')
+const { InvalidParameter } = require('..')
+const GithubDirectoryFileCount = require('./github-directory-file-count.service')
+
+describe('GithubDirectoryFileCount', function () {
+  const contents = [
+    { path: 'a', type: 'dir' },
+    { path: 'b', type: 'dir' },
+    { path: 'c.js', type: 'file' },
+    { path: 'd.js', type: 'file' },
+    { path: 'e.txt', type: 'file' },
+    { path: 'f', type: 'submodule' },
+  ]
+
+  test(GithubDirectoryFileCount.transform, () => {
+    given(contents, {
+      type: 'dir',
+    }).expect({
+      count: 2,
+    })
+    given(contents, {
+      type: 'file',
+    }).expect({
+      count: 3,
+    })
+    given(contents, {
+      type: 'file',
+      extension: 'js',
+    }).expect({
+      count: 2,
+    })
+  })
+
+  it('throws InvalidParameter on receving an object as contents instead of an array', function () {
+    expect(() => GithubDirectoryFileCount.transform({}, {}))
+      .to.throw(InvalidParameter)
+      .with.property('prettyMessage', 'not a directory')
+  })
+})
